feat: #241 출석뷰 연결
