+++ Spoony-iOS/Spoony-iOS/Source/Feature/MyPage/MyPageView.swift
@@ -29,23 +29,26 @@ struct MyPageView: View {
     }
     
     var body: some View {
-            TCARouter(store.scope(state: \.routes, action: \.router)) { screen in
-                switch screen.case {
-                case let .profile(store):
-                    ProfileView(store: store)
-                case let .reviews(store):
-                    ReviewsView(store: store)
-                case let .following(store):
-                    FollowingView(store: store)
-                case let .follower(store):
-                    FollowerView(store: store)
-                case let .editProfile(store):
-                    EditProfileView(store: store)
-                case let .settings(store):
-                    SettingsView(store: store)
-                case let .attendance(store):
-                    AttendanceView(store: store)
-                }
+        TCARouter(store.scope(state: \.routes, action: \.router)) { screen in
+            switch screen.case {
+            case let .profile(store):
+                ProfileView(store: store)
+            case let .reviews(store):
+                ReviewsView(store: store)
+            case let .following(store):
+                FollowingView(store: store)
+            case let .follower(store):
+                FollowerView(store: store)
+            case let .editProfile(store):
+                EditProfileView(store: store)
+            case let .settings(store):
+                SettingsView(store: store)
+            case let .attendance(store):
+                MealTrackerView()
+                    .onDisappear {
+                        store.send(.routeToPreviousScreen)
+                    }
             }
         }
     }
+}
